* Move MatrixDomain tests to new test-matrix-domain.h * Create new benchmark-matrix-domain to test the speed of MatrixDomain-operations * Compilation-fixes to SubvectorFactory<DenseMatrix> * Update GF2 to new interface with zero, one, minusOne * Fix output of field-elements for Modular<uint8> * Add parametrised copy-constructor for Subvector, DenseMatrixRowIterator

+++ linbox/field/modular.h
@@ -218,8 +218,7 @@ public:
 	 * @param  x   field base element.
 	 */
 	std::ostream &write (std::ostream &os, const Element &x) const
-		{ return os << x; }
- 
+		{ return os << (long) x; }
 
 	/*- Read field base element.
 	 * This function assumes the field base element has already been
